+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/graph/model/ForItem.java
@@ -171,4 +171,14 @@ public class ForItem extends LoopItem implements Block {
     public GraphTargetItem returnType() {
         return TypeItem.UNBOUNDED;
     }
+
+    @Override
+    public boolean hasBaseBody() {
+        return true;
+    }
+
+    @Override
+    public List<GraphTargetItem> getBaseBodyCommands() {
+        return commands;
+    }
 }