+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3ClassicAirDecompileTest.java
@@ -1899,8 +1899,8 @@ public class ActionScript3ClassicAirDecompileTest extends ActionScript3Decompile
                 + "}\r\n"
                 + "catch(e:Error)\r\n"
                 + "{\r\n"
-                + "return 4;\r\n"
-                + "}\r\n",
+                + "}\r\n"
+                + "return 4;\r\n",
                  false);
     }
 
@@ -2138,11 +2138,9 @@ public class ActionScript3ClassicAirDecompileTest extends ActionScript3Decompile
                 + "}\r\n"
                 + "catch(e:EOFError)\r\n"
                 + "{\r\n"
-                + "continue;\r\n"
                 + "}\r\n"
                 + "catch(e:Error)\r\n"
                 + "{\r\n"
-                + "continue;\r\n"
                 + "}\r\n"
                 + "}\r\n",
                  false);