+++ CHANGELOG.md
@@ -34,6 +34,8 @@ All notable changes to this project will be documented in this file.
 - [#2077] AS3 direct editation - incorrect slot names
 - [#2077] AS3 direct editation - switch statement missing offset
 - [#2077] AS3 allow star string as property name (XML access)
+- [#2077] AS3 try..catch parts outside block
+- AS3 try..catch inside loop unneccessary continue
 
 ### Changed
 - [#2070] String values inside SWF to XML export are backslash escaped